+++ spring-cloud-gray-webui/src/views/gray-decision/complex-table.vue
@@ -27,24 +27,34 @@
           <span>{{ scope.row.id }}</span>
         </template>
       </el-table-column>
-      <el-table-column label="Name" width="110px" align="center">
+      <el-table-column label="Name" align="center">
         <template slot-scope="scope">
           <span>{{ scope.row.name }}</span>
         </template>
       </el-table-column>
-      <el-table-column label="Infos" width="180px" align="center">
+      <el-table-column label="Infos" align="center">
         <template slot-scope="scope">
           <span>{{ scope.row.infos }}</span>
         </template>
       </el-table-column>
-      <el-table-column label="Policy Id" width="110px" align="center">
+      <el-table-column label="Policy Id" align="center">
         <template slot-scope="scope">
           <span>{{ scope.row.policyId }}</span>
         </template>
       </el-table-column>
-      <el-table-column label="Instance Id" width="110px" align="center">
+      <!--<el-table-column label="Instance Id" align="center">
         <template slot-scope="scope">
           <span>{{ scope.row.instanceId }}</span>
+        </template>
+      </el-table-column>-->
+      <el-table-column label="Operator" prop="operator" align="center">
+        <template slot-scope="scope">
+          <span>{{ scope.row.operator }}</span>
+        </template>
+      </el-table-column>
+      <el-table-column label="Operate Time" prop="operateTime" align="center">
+        <template slot-scope="scope">
+          <span>{{ scope.row.operateTime | parseTime('{y}-{m}-{d} {h}:{i}') }}</span>
         </template>
       </el-table-column>
       <el-table-column label="Actions" align="center" width="230" class-name="small-padding fixed-width">
@@ -168,7 +178,7 @@ export default {
         // Just to simulate the time of the request
         setTimeout(() => {
           this.listLoading = false
-        }, 1.5 * 1000)
+        }, 0.2 * 1000)
       })
     },
     handleFilter() {
@@ -258,20 +268,26 @@ export default {
       })
     },
     handleDelete(row) {
-      deleteDecision(row.id).then(() => {
-        this.dialogFormVisible = false
-        for (const v of this.list) {
-          if (v.id === row.id) {
-            const index = this.list.indexOf(v)
-            this.list.splice(index, 1)
-            break
+      this.$confirm('Confirm to remove the record?', 'Warning', {
+        confirmButtonText: 'Confirm',
+        cancelButtonText: 'Cancel',
+        type: 'warning'
+      }).then(async() => {
+        deleteDecision(row.id).then(() => {
+          this.dialogFormVisible = false
+          for (const v of this.list) {
+            if (v.id === row.id) {
+              const index = this.list.indexOf(v)
+              this.list.splice(index, 1)
+              break
+            }
           }
-        }
-        this.$notify({
-          title: 'Success',
-          message: 'Delete Successfully',
-          type: 'success',
-          duration: 2000
+          this.$notify({
+            title: 'Success',
+            message: 'Delete Successfully',
+            type: 'success',
+            duration: 2000
+          })
         })
       })
     },
